treewide: Fix float to double promotion error
Implicit promotion from float to double type now results in a
compilation warning, treated by CI as an error, hence use
explicit type cast wherever applicable.

+++ applications/asset_tracker_v2/src/modules/modem_module.c
@@ -212,7 +212,7 @@ static void lte_evt_handler(const struct lte_lc_evt *const evt)
 
 		len = snprintf(log_buf, sizeof(log_buf),
 			       "eDRX parameter update: eDRX: %.2f, PTW: %.2f",
-			       evt->edrx_cfg.edrx, evt->edrx_cfg.ptw);
+			       (double)evt->edrx_cfg.edrx, (double)evt->edrx_cfg.ptw);
 		if (len > 0) {
 			LOG_DBG("%s", log_buf);
 		}
